feat: implement new layout and success pages for event photo requests, update image assets
